[Contracts] Rename ServiceLocatorTest
